Provide Request::getUploadedFiles() implementation
