+++ Slim/Http/Body.php
@@ -8,10 +8,6 @@
  */
 namespace Slim\Http;
 
-use InvalidArgumentException;
-use RuntimeException;
-use Psr\Http\Message\StreamInterface;
-
 /**
  * Body
  *
@@ -20,362 +16,7 @@ use Psr\Http\Message\StreamInterface;
  *
  * @link https://github.com/php-fig/http-message/blob/master/src/StreamInterface.php
  */
-class Body implements StreamInterface
+class Body extends Stream
 {
-    /**
-     * Resource modes
-     *
-     * @var  array
-     * @link http://php.net/manual/function.fopen.php
-     */
-    protected static $modes = [
-        'readable' => ['r', 'r+', 'w+', 'a+', 'x+', 'c+'],
-        'writable' => ['r+', 'w', 'w+', 'a', 'a+', 'x', 'x+', 'c', 'c+'],
-    ];
 
-    /**
-     * The underlying stream resource
-     *
-     * @var resource
-     */
-    protected $stream;
-
-    /**
-     * Stream metadata
-     *
-     * @var array
-     */
-    protected $meta;
-
-    /**
-     * Is this stream readable?
-     *
-     * @var bool
-     */
-    protected $readable;
-
-    /**
-     * Is this stream writable?
-     *
-     * @var bool
-     */
-    protected $writable;
-
-    /**
-     * Is this stream seekable?
-     *
-     * @var bool
-     */
-    protected $seekable;
-
-    /**
-     * The size of the stream if known
-     *
-     * @var null|int
-     */
-    protected $size;
-
-    /**
-     * Create a new HTTP message body.
-     *
-     * @param  resource                  $stream A PHP resource handle.
-     * @throws \InvalidArgumentException If argument is not a resource.
-     */
-    public function __construct($stream)
-    {
-        $this->attach($stream);
-    }
-
-    /**
-     * Get stream metadata as an associative array or retrieve a specific key.
-     *
-     * The keys returned are identical to the keys returned from PHP's
-     * stream_get_meta_data() function.
-     *
-     * @link http://php.net/manual/en/function.stream-get-meta-data.php
-     * @param string $key Specific metadata to retrieve.
-     * @return array|mixed|null Returns an associative array if no key is
-     *     provided. Returns a specific key value if a key is provided and the
-     *     value is found, or null if the key is not found.
-     */
-    public function getMetadata($key = null)
-    {
-        if (is_null($key) === true) {
-            return $this->meta;
-        }
-
-        return isset($this->meta[$key]) ? $this->meta[$key] : null;
-    }
-
-    /**
-     * Is a resource attached to this HTTP message body?
-     *
-     * Note: This method is not part of the PSR-7 standard.
-     *
-     * @return bool
-     */
-    public function isAttached()
-    {
-        return is_resource($this->stream);
-    }
-
-    /**
-     * Attach new resource to this HTTP message body.
-     *
-     * Note: This method is not part of the PSR-7 standard.
-     *
-     * @param  resource                  $newStream A PHP resource handle.
-     * @throws \InvalidArgumentException If argument is not a valid PHP resource.
-     */
-    public function attach($newStream)
-    {
-        if (is_resource($newStream) === false) {
-            throw new InvalidArgumentException('\Slim\Http\Body::attach() argument must be a valid PHP resource');
-        }
-
-        if ($this->isAttached() === true) {
-            $this->detach();
-        }
-
-        $this->stream = $newStream;
-        $this->meta = stream_get_meta_data($newStream);
-    }
-
-    /**
-     * Separates any underlying resources from the stream.
-     *
-     * After the stream has been detached, the stream is in an unusable state.
-     *
-     * @return resource|null Underlying PHP stream, if any
-     */
-    public function detach()
-    {
-        $oldResource = $this->stream;
-        $this->stream = null;
-        $this->meta = null;
-        $this->readable = null;
-        $this->writable = null;
-        $this->seekable = null;
-        $this->size = null;
-
-        return $oldResource;
-    }
-
-    /**
-     * Reads all data from the stream into a string, from the beginning to end.
-     *
-     * This method MUST attempt to seek to the beginning of the stream before
-     * reading data and read the stream until the end is reached.
-     *
-     * Warning: This could attempt to load a large amount of data into memory.
-     *
-     * This method MUST NOT raise an exception in order to conform with PHP's
-     * string casting operations.
-     *
-     * @see http://php.net/manual/en/language.oop5.magic.php#object.tostring
-     * @return string
-     */
-    public function __toString()
-    {
-        return $this->isAttached() ? (string) stream_get_contents($this->stream, -1, 0) : '';
-    }
-
-    /**
-     * Closes the stream and any underlying resources.
-     *
-     * @return void
-     */
-    public function close()
-    {
-        if ($this->isAttached() === true) {
-            fclose($this->stream);
-        }
-
-        $this->detach();
-    }
-
-    /**
-     * Get the size of the stream if known.
-     *
-     * @return int|null Returns the size in bytes if known, or null if unknown.
-     */
-    public function getSize()
-    {
-        if (!$this->size && $this->isAttached() === true) {
-            $stats = fstat($this->stream);
-            $this->size = isset($stats['size']) ? $stats['size'] : null;
-        }
-
-        return $this->size;
-    }
-
-    /**
-     * Returns the current position of the file read/write pointer
-     *
-     * @return int Position of the file pointer
-     * @throws \RuntimeException on error.
-     */
-    public function tell()
-    {
-        if (!$this->isAttached() || ($position = ftell($this->stream)) === false) {
-            throw new RuntimeException('Could not get the position of the pointer in stream');
-        }
-        return $position;
-    }
-
-    /**
-     * Returns true if the stream is at the end of the stream.
-     *
-     * @return bool
-     */
-    public function eof()
-    {
-        return $this->isAttached() ? feof($this->stream) : true;
-    }
-
-    /**
-     * Returns whether or not the stream is readable.
-     *
-     * @return bool
-     */
-    public function isReadable()
-    {
-        if ($this->readable === null) {
-            $this->readable = false;
-            if ($this->isAttached()) {
-                foreach (self::$modes['readable'] as $mode) {
-                    if (strpos($this->meta['mode'], $mode) === 0) {
-                        $this->readable = true;
-                        break;
-                    }
-                }
-            }
-        }
-
-        return $this->readable;
-    }
-
-    /**
-     * Returns whether or not the stream is writable.
-     *
-     * @return bool
-     */
-    public function isWritable()
-    {
-        if ($this->writable === null) {
-            $this->writable = false;
-            if ($this->isAttached()) {
-                foreach (self::$modes['writable'] as $mode) {
-                    if (strpos($this->meta['mode'], $mode) === 0) {
-                        $this->writable = true;
-                        break;
-                    }
-                }
-            }
-        }
-
-        return $this->writable;
-    }
-
-    /**
-     * Returns whether or not the stream is seekable.
-     *
-     * @return bool
-     */
-    public function isSeekable()
-    {
-        if ($this->seekable === null) {
-            $this->seekable = false;
-            if ($this->isAttached()) {
-                $this->seekable = $this->meta['seekable'];
-            }
-        }
-
-        return $this->seekable;
-    }
-
-    /**
-     * Seek to a position in the stream.
-     *
-     * @link http://www.php.net/manual/en/function.fseek.php
-     * @param int $offset Stream offset
-     * @param int $whence Specifies how the cursor position will be calculated
-     *     based on the seek offset. Valid values are identical to the built-in
-     *     PHP $whence values for `fseek()`.  SEEK_SET: Set position equal to
-     *     offset bytes SEEK_CUR: Set position to current location plus offset
-     *     SEEK_END: Set position to end-of-stream plus offset.
-     * @throws \RuntimeException on failure.
-     */
-    public function seek($offset, $whence = SEEK_SET)
-    {
-        // Note that fseek returns 0 on success!
-        if (!$this->isSeekable() || fseek($this->stream, $offset, $whence) === -1) {
-            throw new RuntimeException('Could not to seek in stream');
-        }
-    }
-
-    /**
-     * Seek to the beginning of the stream.
-     *
-     * If the stream is not seekable, this method will raise an exception;
-     * otherwise, it will perform a seek(0).
-     *
-     * @see seek()
-     * @link http://www.php.net/manual/en/function.fseek.php
-     * @throws \RuntimeException on failure.
-     */
-    public function rewind()
-    {
-        if (!$this->isSeekable() || rewind($this->stream) === false) {
-            throw new RuntimeException('Could not to rewind stream');
-        }
-    }
-
-    /**
-     * Read data from the stream.
-     *
-     * @param int $length Read up to $length bytes from the object and return
-     *     them. Fewer than $length bytes may be returned if underlying stream
-     *     call returns fewer bytes.
-     * @return string Returns the data read from the stream, or an empty string
-     *     if no bytes are available.
-     * @throws \RuntimeException if an error occurs.
-     */
-    public function read($length)
-    {
-        if (!$this->isReadable() || ($data = fread($this->stream, $length)) === false) {
-            throw new RuntimeException('Could not read from stream');
-        }
-        return $data;
-    }
-
-    /**
-     * Write data to the stream.
-     *
-     * @param string $string The string that is to be written.
-     * @return int Returns the number of bytes written to the stream.
-     * @throws \RuntimeException on failure.
-     */
-    public function write($string)
-    {
-        if (!$this->isWritable() || ($written = fwrite($this->stream, $string)) === false) {
-            throw new RuntimeException('Could not write to stream');
-        }
-        return $written;
-    }
-
-    /**
-     * Returns the remaining contents in a string
-     *
-     * @return string
-     * @throws \RuntimeException if unable to read or an error occurs while
-     *     reading.
-     */
-    public function getContents()
-    {
-        if (!$this->isReadable() || ($contents = stream_get_contents($this->stream)) === false) {
-            throw new RuntimeException('Could get contents of stream');
-        }
-        return $contents;
-    }
 }
